Refactor StepFilterTests
